AMR Service: Fix full image build error.
Delete header files.
