+++ qrb_ros_amr/include/action/cmd_action_server.hpp
@@ -7,6 +7,7 @@
 #define QRB_ROS_AMR__CMD_ACTION_SERVER_HPP_
 
 #include "amr_manager.hpp"
+#include "ros_common.hpp"
 
 using namespace qrb::amr_manager;
