+++ qrb_ros_amr/include/amr_manager/ros_common.hpp
@@ -75,4 +75,44 @@ using CommandCode = qrb_ros_slam_msgs::msg::CommandCode;
 
 #define SERVICE_TIMEOUT_DURATION 5  // timeout is 5 seconds
 
+class API
+{
+public:
+  const static int INIT_AMR = 1;
+  const static int RELEASE_AMR = 2;
+  const static int ENABLE_DEVELOPER_MODE = 11;
+  const static int ME_COMPLETED = 12;
+
+  static std::string to_string(int cmd)
+  {
+    std::string message;
+    switch (cmd) {
+      case INIT_AMR:
+        message = "INIT_AMR";
+        break;
+      case RELEASE_AMR:
+        message = "RELEASE_AMR";
+        break;
+      case ENABLE_DEVELOPER_MODE:
+        message = "ENABLE_DEVELOPER_MODE";
+        break;
+      case ME_COMPLETED:
+        message = "ME_COMPLETED";
+        break;
+      default:
+        message = "INVAILD API";
+        break;
+    }
+    return message;
+  }
+};
+
+enum class Lifecycle_State
+{
+  UnConfigured = 0,
+  Inactive = 1,
+  Active = 2,
+  Finalized = 3,
+};
+
 #endif  // QRB_ROS_AMR__ROS_COMMON_HPP_
