+++ qrb_ros_amr/include/service/api_service_server.hpp
@@ -8,6 +8,7 @@
 
 #include "developer_mode_subscriber.hpp"
 #include "amr_manager.hpp"
+#include "ros_common.hpp"
 
 using namespace qrb::amr_manager;
