+++ qrb_ros_amr/include/service/cartographer_service_client.hpp
@@ -7,6 +7,7 @@
 #define QRB_ROS_AMR__CARTOGRAPHER_SERVICE_CLIENT_HPP_
 
 #include "amr_manager.hpp"
+#include "ros_common.hpp"
 
 using namespace qrb::amr_manager;
