+++ qrb_ros_amr/include/topic/map_subscriber.hpp
@@ -8,6 +8,7 @@
 
 #include "amr_status_transporter.hpp"
 #include "amr_manager.hpp"
+#include "ros_common.hpp"
 
 using namespace qrb::amr_manager;
